pyproject.toml
[tool.poetry]
name = "favorite_movies"
version = "0.1.0"
description = ""
authors = ["UchavtovAV"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.13.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-multipart = "^0.0.17"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
